Fix: use MixOptimizer instead of RiemannianAdam because SPDNet is defined for MixOptimizer
